Get to the point where it almost starts to work
eventlet spawns the scgi server which is able to create a new console session.
So far, not actually hooked in any way to a backend.
The plan will be to hook up read and write of ConsoleSession instance handle method.

That handle will in turn be hooked into the logging/watching scheme which in turn is
hooked into the likes of pyghmi.ipmi.console and/or ssh execution