Fix json output mistakes.
I may wish to change http api to a single json dumps.  At that point the
socket api would be the only consumer of the piecewise json messages.
Merging the json data is a non-trivial problem and httpapi is doing too much
hand assembly of the data.

http api would lose any hope of one request, staggered response behavior as a result,
but http is a terrible protocol for that anyway and there is always the socket api