Rework ipmi console object to block on connect
Previously, the state would be seen as 'connected' and then 'disconnected' in event of
connection failing.  Rework things such that the console session stays in 'connecting' state
until timeout or success occurs and don't send disconnect, instead raising an exception.
This makes the connection action a bit more intuitive to the user, who would assume a 'connected'
console means the endpoint was reachable.  This may not always be possible in a console plugin,
but it's a nice pattern when possible.  If a console plugin cannot tell when 'connected' happens, then
the previous behavior of this plugin makes sense as a 'best effort': return 'connected', send
disconnect event when the console turns out to be bad.  For example, executable consoles are most
likely going to follow this pattern.  An option could be for an executable to have a certain
signature to print to show 'connected' though...