Removing the confluent_xcat subdirectory
After some discussion with the team, let's keep the xcat
related "glue" out of the confluent project, for now.
This will give me more time to figure out how things will
come together and integrate for xCAT3.

For xCAT2, a xCAT-confluent RPM is added to xcat-core project to
be consistent with how xCAT2 is packaged today.