Begin work on node handlers
The handler framework will be the key for supporting multiple node types.
Each handler is supposed to support a variety of stages through the
process.

First a unicast interrogation to get some data (ideally without
logging in, or by logging in with default credentials).

Second upon identification, preconfig will be called, if and only if the endpoint would become a node,
but should not yet do anything big, just paving the way for verification or dependent discovery configuration.

Finally, config will be called to deploy configuration according to the node and referenced
config manager.