Further advance discovery
This commit starts to lay the groundwork for
handling the security policy differences with
respect to TLS keys on port 443.  This requires
that the discovered entitity support https for now,
which is a given (even if the web service is not enabled,
it does provide a cert).

IMMv2 may never be able to follow the 'secure' path due
to uncertainty of whether there's a licensed web ui or not).