Fix discovery and messaging around certain scenarios
If a node has a random ip address, then static, then goes back,
it was not updating the addresses.  When an address is updated,
go ahead and do a deeper investigation.

If anything should go wrong unexpectedly, don't let the periodic recheck
forever die.

If for some reason we have bad ip in the info dictionary, log a more accurate
message.