Numerous discovery anomaly fixes
For one, ensure mac address is valid in the neighutil.py, rather than
trying to blacklist.  Apart from FAILED, INCOMPLETE is possible,
and it's just a bad idea to blacklist.

When getting the https_cert has an unexpected situation, now
say it is because of unreachable all the time.

Refactor the loop in the node recheck so that the inner loop is brought out.
Catch and print any exceptions, but carry on to the next entry regardless.

Do not exit out of above loop because one thing is unreachable.