Implement no-qbridge mode and workaround for switch mistake
Sometimes switches erroneously use interface index instead of bridge index
in the bridge mib.  Workaround by assuming this is probably the case if
none of the macs seem to map to an interface.  Note that it's still important
to accept that a bridge port might not have an interface (internal mac address)
so only make the assumption if *none* of the mac addresses mapped to an interface.