Allow memory reclamation through deletion
When going to zram, things were solid for space reduction as
data was written, however memory could no longer be reclaimed.

It turns out that zram supports TRIM, and by telling xfs discard,
we have it do trim-on-demand. It is by default off out of performance
concerns, but I don't think that applies to a zram backed filesystem.